+++ kroto-plus-coroutines/src/main/kotlin/com/github/marcoferrer/krotoplus/coroutines/server/ServerCalls.kt
@@ -139,6 +139,8 @@ public fun <ReqT, RespT> ServiceScope.serverCallBidiStreaming(
 
     val responseChannel = Channel<RespT>()
     val serverCallObserver = (responseObserver as ServerCallStreamObserver<RespT>)
+        .apply { disableAutoInboundFlowControl() }
+
     with(newRpcScope(initialContext, methodDescriptor)) rpcScope@ {
         bindToClientCancellation(serverCallObserver)
         applyOutboundFlowControl(serverCallObserver,responseChannel)